+++ lib/deas/sinatra_runner.rb
@@ -1,9 +1,9 @@
-require 'deas/runner'
+require 'deas/deas_runner'
 require 'deas/template'
 
 module Deas
 
-  class SinatraRunner < Runner
+  class SinatraRunner < DeasRunner
 
     def initialize(handler_class, sinatra_call)
       @sinatra_call = sinatra_call
@@ -11,19 +11,11 @@ module Deas
       super(handler_class, {
         :request  => @sinatra_call.request,
         :response => @sinatra_call.response,
-        :params   => NormalizedParams.new(@sinatra_call.params).value,
+        :params   => @sinatra_call.params,
         :logger   => @sinatra_call.settings.logger,
         :router   => @sinatra_call.settings.router,
         :session  => @sinatra_call.session,
       })
-    end
-
-    def run
-      run_callbacks @handler_class.before_callbacks
-      @handler.init
-      response_data = @handler.run
-      run_callbacks @handler_class.after_callbacks
-      response_data
     end
 
     # Helpers
@@ -74,18 +66,8 @@ module Deas
 
     private
 
-    def run_callbacks(callbacks)
-      callbacks.each{|proc| @handler.instance_eval(&proc) }
-    end
-
     def get_content_type(template_name)
       File.extname(template_name)[1..-1] || 'html'
-    end
-
-    class NormalizedParams < Deas::Runner::NormalizedParams
-      def file_type?(value)
-        value.kind_of?(::Tempfile)
-      end
     end
 
   end
